Refactor the tool interface (again) to allow multi-page tools, use message boxes for alerts where applicable
All that's left is to add default value support, which is just an extension, and then the tool interface ought to be final-ready.

This is for section editing:
edit-section (selector) -> edit-section-x -> null

(edit-section-x needs to be a separate tool because of the complete change between different sections.)

After that, EMI's GUI is actually feature-complete - the rest is backend work.